Bugfixes and improvements keeping up with python3 + ionize.

+++ emigrate/flux_schemes/Differentiate.py
@@ -96,8 +96,8 @@ class Differentiate(object):
                                   [99./80., -3., 93./40., -3./5., 3./80]]
             # note typo in paper saying -3/80
         elif self.method == 'dissipative':
-            internal_function = [1, -2, 1]
-            boundary_functions = [[1, -2, 1]]
+            internal_function = [1., -2., 1.]
+            boundary_functions = [[1., -2., 1.]]
 
         self.B2 = self.construct_matrix(boundary_functions,
                                         internal_function,
@@ -118,7 +118,7 @@ class Differentiate(object):
         l = len(internal_function)
         construct = [[0]*i + internal_function +
                      [0] * (N-i+1) for i in range(N)]
-        construct = np.array(construct)[:, (l-1.)/2.:-(l+3.)/2.]
+        construct = np.array(construct)[:, int((l-1.)/2.):int(-(l+3.)/2.)]
 
         for idx, func in enumerate(boundary_functions):
             construct[idx, :] = func + [0] * (N - len(func))
